Modify DDRA diagnostics to work with long long's = int64 sizes

+++ src/nco/nco_ctl.h
@@ -1,4 +1,4 @@
-/* $Header: /data/zender/nco_20150216/nco/src/nco/nco_ctl.h,v 1.21 2006-04-30 06:18:26 zender Exp $ */
+/* $Header: /data/zender/nco_20150216/nco/src/nco/nco_ctl.h,v 1.22 2006-04-30 21:13:22 zender Exp $ */
 
 /* Purpose: Program flow control functions */
 
@@ -66,9 +66,9 @@ nco_ddra /* [fnc] Count operations */
  const int rnk_wgt, /* I [nbr] Rank of weight */
  const int var_idx, /* I [enm] Index */
  const int wrd_sz, /* I [B] Bytes per element */
- const long lmn_nbr, /* I [nbr] Variable size */
- const long lmn_nbr_avg, /* I [nbr] Averaging block size */
- const long lmn_nbr_wgt); /* I [nbr] Weight size */
+ const long long lmn_nbr, /* I [nbr] Variable size */
+ const long long lmn_nbr_avg, /* I [nbr] Averaging block size */
+ const long long lmn_nbr_wgt); /* I [nbr] Weight size */
 
 void
 nco_lbr_vrs_prn(void); /* [fnc] Print netCDF library version */
